requirements.txt
requests==2.31.0
python-dotenv==1.0.0
chromadb==0.4.24
sentence-transformers==2.7.0